refactor: split decisions.rs (3074 lines) into 3 modules
Pure refactor, zero behavior changes. All existing import paths
preserved via re-exports from decisions.rs.

- decisions.rs (992 lines) — records, types, logging, JSONL I/O,
  stats, paths, background distillation trigger, re-exports
- preferences.rs (1950 lines) — PreferencePattern, DistilledPreferences,
  distillation engine (Gini impurity, temporal patterns, outcome
  backfilling), preference save/load, format_preference_summary
- retrieval.rs (302 lines) — few-shot retrieval, similarity scoring,
  format_few_shot_examples
