+++ src/brain/mod.rs
@@ -7,4 +7,6 @@ pub mod evals;
 pub mod insights;
 pub mod mailbox;
 pub mod metrics;
+pub mod preferences;
 pub mod prompts;
+pub mod retrieval;
